requirements.txt
Django==1.9.1
djangorestframework==3.3.2
httpie==0.9.3
Pygments==2.0.2
requests==2.9.1
wheel==0.24.0
